A new concept from faraway Earth: Verbs!
I've finished the first draft of my new dictionary. I added more words for antonyms, since I don't intend to have a `iat` equivalent this time around, though I may consider one for the technical nouns.
I also added verbs. They don't need prepositions anymore because their presence in the sentence has one meaning: the predicate. Most of the things I once called nouns are now verbs. Nouns are just the handful of things that I couldn't define with verbs.
I also modified the word generator to reflect my new treatment of `l` as a vowel. It appears as the second vowel in dipthongs and as an approximant in special cases, but never as a standalone vowel.