There's such a thing as "too many phonemes"
I reduced the number of allowable consonant clusters and reduced the number of words I generated to improve word distinctiveness. I was just generating such an absurdly larger number of words with such absurdly loose syllable structure rules that I was getting basically a buch of 2- and 3-phoneme words that were hard to pronounce. So I scaled back to reflect the limited size of my vocabulary.