Aw, frick (ates)!
I made some modifications to my phoneme analysis code and added a new phoneme to my list. This was due to my realisation that while [ʃ] and [t̠ʃ] are not very common, and [ʃ] and [ʂ] are not very common, and [ʂ] and [ʈʂ] are not very common, [ʃ], [ʂ], [t̠ʃ], and [ʈʂ] put together are more common than [w], and just common enough for me to slide it into my inventory while maintaining a solid majority of humans at 56.5%. I came to this realisation thanks to one of Jan Misali's comments on one of his own videos. This brings my set back up to 16 phonemes, will put the letter "c" back in my language, and, to my exitement, gives me an affricate with which to work again! I like affricates. I also came up with a slightly more intelligent way to deal with aspiration.