Word order is for squares!
I've had what I believe to be a stroke of genius. There are no adjectives. All relative clauses are headless. A noun phrase is a list of nouns, and refers to their intersection! The modifiers and nouns can come in any order because their all nouns! Well, a few things still need to be qualifiers and have a directionality to them, but I have moved many "qualifier"s over to "pronoun".
I also adjusted my language selection algorithm, but my sources are the same.