Don't pepper my zision!
I finally got around to filling out my prepositions. That was the biggest particle class, if I'm not mistaken, so I'm glad to have it set. I was able to get a good diversity of accurate, recognisable source words if I do say so myself.
I also tentatively added a new class of words, abbreviations. These are relative clauses of more complex verbs that are not themselves words. I'm not sure if this is the best way to do this, but I'm going with it for now, at least for note-taking purposes.