The Golden Star of Shastisu!
The flag of Shastisu contains a horizontal sinusoidal bicolor of white and turquoise. The sinusoid completes one period across the banner, centered at the edges and middle, higher on the hoist half, and lower on the fly half. The charge is a circle divided into equal parts turquoise and white with an orange sixpointed star on it. The white and turquoise represent peace and knowledge, the turquoise and white circle represents the Earth, and the star represents the glory of the Shastisu language, connecting the six continents of the Earth.