Pauli's Inclusion
I changed my pronouns back to a clusive system rather than a numeric system. It's not quite as common, but it doesn't make sense to have just plural pronouns when none of my other words inflect for number; inflecting for clusivity makes much more sense, as all pronouns inflect for person, which is basically the same thing.