Better by the minute
Lots of big changes. I have committed to a self-segregating morphology and changed what few words I have accordingly. I added numerals for a lot and a little. I changed a bunch of base words to compounds. I added a bunch of mnemonic opposites. I merged my distal demonstrative with my third person pronoun. I changed stimuli to be themes and experiencers to be patients. And I made progress on the dictionary.