9 vowels... 5 graphemes... wat
I did the broad transcriptions of Sotho. It wasn't too challenging, since I was able to get a digraph and trigraph system that made most of it pretty simple, though syllabics posed a small challenge as did the fact that all three mid-front vowels are written <e>, and same for the backs. I just guessed /e/ and /o/ for all of them, and then defaulted to a list of the examples I saw on the Wikipedia page for all others.