Shedding morphemes like it's going out of style
I've been working on the vocab. The lexicon of morphemes gets smaller and smaller, thank goodness. At this rate, I should have no problem getting well below Votgil's 600 roots. I also made the bold choice to drop the word for "not"; the word "be false" shall be used as an auxiliary to mark polarity. I still need a word for "even", so those can't be antonyms anymore, but I think this is the right move.