Count it properly
I fixed the way it does cessatives to use antonyms, not "nos", I made the thing to tell if something needs an antonym smarter, I made sure it loads onomotopoeias and takes them into account when choosing words, and added a thing where it counts the results at the end. Also minor improvements to the Esperanto parsing.