Shifting meanings
It turns out that GT doesn't actually support direct translation between any pair of languages that doesn't include English. So using es:"mosca" rather than en:"fly" didn't actually do anything to clarify that I wanted the noun, not the verb, because it just translated it into English first. So I had to go rethink how I'm doing those tricky meanings. I think the answer is a lot of online dictionaries. It's slower than GT, as none are as extensive or as well supported, but I can do it for just the tricky ones and then the set of words that actually gets chosen.