Catt-catt
I straightened out the english and spanish transcriptions. On the english side, it was mostly clarifying /ɑ/s /ɒ/s and /ɔ/s (engtoipa was strangely inconsistent on that front), adding words it didn't know, correcting allophones gone awry, and changing broad transcriptions of schwas to their historic values, to give more recognisable words. On the spanish side, I changed a bunch of /we/s to /o/s, to make it more similar to other Latin languages (but only where conjugation might cause it to be /o/ in modern Spanish; I don't want to go too deep off the etymological... rail? path?)