A mayor yump
In addition to adding more endonyms and fleshing out more national identities, I made the decision to change /j/ from ⟨j⟩ to ⟨y⟩, to conform to the vast majority of languages that use the latin script. It will make my alphabet slightly less compatible with IPA, but I think that's fine; I'll just use ⟨u̟⟩ instead of ⟨y⟩. This was an imporant decision in making my language truly neutral.