No, I did NOT mean Japanese
I watched so many terrible YouTube videos and sifted through so many Japanese-related results to do this one.
I realised that Google wasn't giving me the accents over the es in Javanese, which actually proves the difference between schwa and front mid vowel. It was crazy how hard it was to find out what the difference between acute and gave accents are; apparently it's allophonic and not actually important.