My first stylistic choice!
I noticed a bunch of things as I began translating things into Oltilip for the first time. The biggest issue I noticed was that the word for "be part of", onco, being the most common postposition plus the word for "mean", co, was super confusing in compound words involving the both of them. For example, colononcolon: part of speech. Like, what?
Also a bunch of minor improvements across the board.