I was terrible
I redid the country names to make sense. Now, if a people group is defined by a country, the root for both is the name of the country. So "kamhun" now, instead of meaning "a Cameroonian" or something dumb like that, means "Cameroon".
Oh, this makes everything so much more logical.
Also, did you know Shanghai isn't pronounced anything like /ʃɑŋˈhɑɪ/. I once had a brief debate with someone about whether people should change to saying it that way. This is a strong argument in my favor.