First draft!
I got the first draft of my Oltilip-to-English dictionary, and it looks great! There are probably some bugs still in it, but so far, I'm extremely pleased. The next step is to add links so that clicking on words takes you to their definitions.

I also just deleted the special word class. I've been neglecting it for ages (one of them still had a <j> in it), and now I think I just don't need them.